renamed comparison files
