+++ .gitignore
@@ -1,3 +1,4 @@
+.ipynb_checkpoints/*
 .vscode/*
 !.vscode/settings.json
 !.vscode/tasks.json
