--- 2017110266/a.js
@@ -1,5 +1,0 @@
-var x = 5.970563;
-
-x = Number(x.toFixed(2));
-console.log(x);
-console.log(typeof(x));
