--- 2017110266/a.py
@@ -1,4 +1,0 @@
-x = 1.232142214214142
-x = float(format(x,'.4f'))
-print(x)
-print(type(x))
