+++ 2017110266/client/LoadTest_Comparison.py
@@ -1,4 +1,4 @@
-import LoadMeasurements
+import LoadTest_Measurements
 import json
 import requests
 
@@ -13,6 +13,6 @@ OS = requests.get('http://localhost:8000/os').content.decode()
 for server in iterations['OS'][OS]['Server']:
     for resource in iterations['OS'][OS]['Server'][server]['resource']:
         for method in iterations['OS'][OS]['Server'][server]['resource'][resource]['method']:
-            results.append(LoadMeasurements.run(server, resource, method))
-            LoadMeasurements.writeResults(OS,server,results,resource,method)
+            results.append(LoadTest_Measurements.run(server, resource, method))
+            LoadTest_Measurements.writeResults(OS,server,results,resource,method)
             results.clear()
